dashboard/requirements.txt
python-dotenv
boto3
dash
pandas